# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransMóvil - Transport & Local Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver public profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverPublic"
        "404":
          description: Not found
...
  /trips:
    post:
      operationId: requestTrip
      tags: [trips]
      summary: Request a new trip (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, driver, admin]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /admin/users/{id}:
    get:
      operationId: getAdminUser
      tags: [admin]
      summary: Get user detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/drivers:
    get:
      operationId: getAdminDrivers
      tags: [admin]
      summary: List all drivers
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, suspended]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Drivers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverListResponse"
  /admin/drivers/{id}:
    patch:
      operationId: updateAdminDriver
      tags: [admin]
      summary: Approve / suspend / update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverProfile"
  /admin/trips:
    get:
      operationId: getAdminTrips
      tags: [admin]
      summary: List all trips
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trips list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripListResponse"
  /admin/config:
    get:
      operationId: getAdminConfig
      tags: [admin]
      summary: Get platform configuration
      responses:
        "200":
          description: Config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
    patch:
      operationId: updateAdminConfig
      tags: [admin]
      summary: Update platform configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformConfigUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConfig"
  /admin/cities:
    get:
      operationId: getAdminCities
      tags: [admin]
      summary: List cities
      responses:
        "200":
          description: Cities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/City"
    post:
      operationId: createCity
      tags: [admin]
      summary: Add a city
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /admin/cities/{id}:
    patch:
      operationId: updateCity
      tags: [admin]
      summary: Update city
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CityInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/City"
  /admin/earnings:
    get:
      operationId: getAdminEarnings
      tags: [admin]
      summary: Platform earnings report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, all]
            default: month
      responses:
        "200":
          description: Earnings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformEarnings"

# ─── SCHEMAS ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, role, firstName, lastName, email, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    DriverPublic:
      type: object
      required: [id, firstName, lastName, rating, totalTrips, isAvailable, serviceTypeId, lat, lng]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    DriverProfile:
      type: object
      required: [id, userId, status, isAvailable, rating, totalTrips, serviceTypeId, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [pending, approved, suspended]
        isAvailable:
          type: boolean
        rating:
          type: number
        totalTrips:
          type: integer
        totalEarnings:
          type: number
        serviceTypeId:
          type: integer
        bio:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        createdAt:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
        vehicle:
          $ref: "#/components/schemas/Vehicle"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
...
    Vehicle:
      type: object
      required: [id, driverId, make, model, year, plate, color, serviceTypeId]
      properties:
        id:
          type: integer
        driverId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        color:
          type: string
        serviceTypeId:
          type: integer
        photoUrl:
          type: ["string", "null"]
        capacity:
          type: integer
        createdAt:
          type: string
...
    Trip:
      type: object
      required: [id, customerId, status, pickupAddress, pickupLat, pickupLng, destinationAddress, destinationLat, destinationLng, serviceTypeId, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [searching, accepted, arriving, in_progress, completed, cancelled]
        pickupAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        serviceTypeId:
          type: integer
        estimatedPrice:
          type: ["number", "null"]
        finalPrice:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/UserProfile"
        driver:
          $ref: "#/components/schemas/DriverPublic"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
    TripRequest:
      type: object
      required: [pickupAddress, pickupLat, pickupLng, destinationAddress, destinationLat, destinationLng, serviceTypeId]
      properties:
        pickupAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        serviceTypeId:
          type: integer
        paymentMethod:
          type: string
          enum: [cash, mercadopago, card]
          default: cash
        notes:
          type: string
...
    ServiceType:
      type: object
      required: [id, name, icon, description, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        basePrice:
          type: number
        pricePerKm:
          type: number
        isActive:
          type: boolean
...
    UserListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        total:
          type: integer
    DriverListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/DriverProfile"
        total:
          type: integer
    TripListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: integer
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [customer, driver, admin]
        isActive:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
    AdminDriverUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, suspended]
        notes:
          type: string
    PlatformConfig:
      type: object
      required: [id, commissionPercent, minTripPrice, maxSearchRadius, allowedPaymentMethods]
      properties:
        id:
          type: integer
        commissionPercent:
          type: number
        minTripPrice:
          type: number
        maxSearchRadius:
          type: number
        allowedPaymentMethods:
          type: array
          items:
            type: string
        maintenanceMode:
          type: boolean
        updatedAt:
          type: string
    PlatformConfigUpdate:
      type: object
      properties:
        commissionPercent:
          type: number
        minTripPrice:
          type: number
        maxSearchRadius:
          type: number
        allowedPaymentMethods:
          type: array
          items:
            type: string
        maintenanceMode:
          type: boolean
    City:
      type: object
      required: [id, name, country, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        state:
          type: ["string", "null"]
        isActive:
          type: boolean
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
    CityInput:
      type: object
      required: [name, country]
      properties:
        name:
          type: string
        country:
          type: string
        state:
          type: string
        isActive:
          type: boolean
        lat:
          type: number
        lng:
          type: number
    PlatformEarnings:
      type: object
      required: [period, totalRevenue, platformRevenue, driverRevenue, totalTrips]
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        platformRevenue:
          type: number
        driverRevenue:
          type: number
        totalTrips:
          type: integer
        topDrivers:
          type: array
          items:
            type: object
            required: [driverId, name, earnings, trips]
            properties:
              driverId:
                type: integer
              name:
                type: string
              earnings:
                type: number
              trips:
                type: integer