# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TransMóvil - Transport & Local Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  driverId: number;
  make: string;
  model: string;
  year: number;
  plate: string;
  color: string;
  serviceTypeId: number;
  /** @nullable */
  photoUrl?: string | null;
  capacity?: number;
  createdAt?: string;
}